Add used market service slide animation
Add used market service slide animation
