+++ koin/src/main/java/in/koreatech/koin/ui/usedmarket/MarketUsedBaseFragment.java
@@ -1,16 +1,17 @@
 package in.koreatech.koin.ui.usedmarket;
 
-import in.koreatech.koin.core.fragment.BaseFragment;
+import androidx.fragment.app.Fragment;
 
-public class MarketUsedBaseFragment extends BaseFragment{
+public class MarketUsedBaseFragment {
 
     /**
      * Change fragment method from TabView
+     *
      * @param page page that selected
      * @return fragment fragment that selected
      */
-    public static MarketUsedBaseFragment newInstance(int page) {
-        MarketUsedBaseFragment fragment = null;
+    public static Fragment newInstance(int page) {
+        Fragment fragment = null;
         switch (page) {
             case 0:
                 fragment = new MarketUsedSellFragment();
