+++ koin/src/main/java/in/koreatech/koin/ui/usedmarket/MarketUsedBuyFragment.java
@@ -3,62 +3,70 @@ package in.koreatech.koin.ui.usedmarket;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
+import com.google.firebase.perf.metrics.AddTrace;
+
+import java.util.ArrayList;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
-
-import com.google.firebase.perf.metrics.AddTrace;
-
 import in.koreatech.koin.R;
-import in.koreatech.koin.ui.usedmarket.presenter.MarketUsedContract;
 import in.koreatech.koin.core.recyclerview.RecyclerClickListener;
 import in.koreatech.koin.core.recyclerview.RecyclerViewClickListener;
 import in.koreatech.koin.core.swiperefreshbottom.SwipeRefreshLayoutBottom;
+import in.koreatech.koin.core.toast.ToastUtil;
 import in.koreatech.koin.data.network.entity.Item;
 import in.koreatech.koin.data.network.interactor.MarketUsedRestInteractor;
 import in.koreatech.koin.data.network.response.MarketPageResponse;
-import in.koreatech.koin.ui.usedmarket.presenter.MarketUsedPresenter;
-import in.koreatech.koin.core.toast.ToastUtil;
+import in.koreatech.koin.ui.koinfragment.KoinBaseFragment;
+import in.koreatech.koin.ui.main.MainActivity;
 import in.koreatech.koin.ui.usedmarket.adapter.MarketUsedBuyRecyclerAdapter;
+import in.koreatech.koin.ui.usedmarket.presenter.MarketUsedContract;
+import in.koreatech.koin.ui.usedmarket.presenter.MarketUsedPresenter;
+import in.koreatech.koin.util.NavigationManger;
 
-import java.util.ArrayList;
-
-public class MarketUsedBuyFragment extends MarketUsedBaseFragment implements MarketUsedContract.View, SwipeRefreshLayoutBottom.OnRefreshListener {
+public class MarketUsedBuyFragment extends KoinBaseFragment implements MarketUsedContract.View, SwipeRefreshLayoutBottom.OnRefreshListener {
     private final String TAG = "MarketUsedBuyFragment";
-
-
+    private final int BUYMARKETID = 1;
+    @BindView(R.id.market_used_buy_recyclerview)
+    RecyclerView marketBuyRecyclerView;
+    @BindView(R.id.market_used_buy_swiperefreshlayout)
+    SwipeRefreshLayoutBottom marketBuySwipeRefreshLayout;
     private Unbinder unbinder;
     private MarketUsedPresenter marketUsedPresenter;
     private ArrayList<Item> marketBuyArrayList = new ArrayList<>(); //DB의 중고장터 리스트를 저장할 ArrayList
-
     /* Adapter */
     private MarketUsedBuyRecyclerAdapter marketUsedBuyRecyclerAdapter;
-
     /* View Component */
     private View view;
     private RecyclerView.LayoutManager layoutManager; // RecyclerView LayoutManager
-
-    private final int BUYMARKETID = 1;
     private int position;
     private boolean grantCheck;
     private int currentPage;
     private int totalPage;
     private boolean isResume;
+    //recyclerview item touch listener
+    private RecyclerClickListener recyclerItemtouchListener = new RecyclerClickListener(getActivity(), marketBuyRecyclerView, new RecyclerViewClickListener() {
+        @Override
+        public void onClick(View view, final int index) {
+            marketUsedPresenter.readGrantedDetail(marketBuyArrayList.get(position).getId());
+            position = index;
+        }
 
-    @BindView(R.id.market_used_buy_recyclerview)
-    RecyclerView marketBuyRecyclerView;
-    @BindView(R.id.market_used_buy_swiperefreshlayout)
-    SwipeRefreshLayoutBottom marketBuySwipeRefreshLayout;
+        @Override
+        public void onLongClick(View view, int position) {
+        }
+
+    });
 
     @Override
     public void onAttach(Context context) {
@@ -83,12 +91,12 @@ public class MarketUsedBuyFragment extends MarketUsedBaseFragment implements Mar
 
     @Override
     public void showLoading() {
-        ((MarketUsedActivity) getActivity()).showProgressDialog(R.string.loading);
+        ((MainActivity) getActivity()).showProgressDialog(R.string.loading);
     }
 
     @Override
     public void hideLoading() {
-        ((MarketUsedActivity) getActivity()).hideProgressDialog();
+        ((MainActivity) getActivity()).hideProgressDialog();
     }
 
     @Override
@@ -104,11 +112,6 @@ public class MarketUsedBuyFragment extends MarketUsedBaseFragment implements Mar
     @Override
     public void onStop() {
         super.onStop();
-    }
-
-    @Override
-    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
     }
 
     @AddTrace(name = "MarketData_Received_Buy")
@@ -146,20 +149,6 @@ public class MarketUsedBuyFragment extends MarketUsedBaseFragment implements Mar
 
         marketUsedBuyRecyclerAdapter.notifyDataSetChanged();
     }
-
-    //recyclerview item touch listener
-    private RecyclerClickListener recyclerItemtouchListener = new RecyclerClickListener(getActivity(), marketBuyRecyclerView, new RecyclerViewClickListener() {
-        @Override
-        public void onClick(View view, final int index) {
-            marketUsedPresenter.readGrantedDetail(marketBuyArrayList.get(position).getId());
-            position = index;
-        }
-
-        @Override
-        public void onLongClick(View view, int position) {
-        }
-
-    });
 
     @Override
     public void onRefresh() {
@@ -220,11 +209,11 @@ public class MarketUsedBuyFragment extends MarketUsedBaseFragment implements Mar
     @Override
     public void onGrantedDataReceived(boolean granted) {
         this.grantCheck = granted;
-        Intent intent = new Intent(getActivity(), MarketUsedBuyDetailActivity.class);
-        intent.putExtra("ITEM_ID", marketBuyArrayList.get(this.position).getId());
-        intent.putExtra("MARKET_ID", BUYMARKETID);
-        intent.putExtra("GRANT_CHECK", this.grantCheck);
-        startActivity(intent);
+        Bundle bundle = new Bundle();
+        bundle.putInt("ITEM_ID", marketBuyArrayList.get(this.position).getId());
+        bundle.putInt("MARKET_ID", BUYMARKETID);
+        bundle.putBoolean("GRANT_CHECK", this.grantCheck);
+        NavigationManger.getNavigationController(getActivity()).navigate(R.id.navi_market_used_buy_detail_action, bundle, NavigationManger.getNavigationAnimation());
     }
 
     @Override
