+++ koin/src/main/java/in/koreatech/koin/ui/usedmarket/MarketUsedSellFragment.java
@@ -3,62 +3,71 @@ package in.koreatech.koin.ui.usedmarket;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
+import com.google.firebase.perf.metrics.AddTrace;
+
+import java.util.ArrayList;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
-
-import com.google.firebase.perf.metrics.AddTrace;
-
 import in.koreatech.koin.R;
-import in.koreatech.koin.ui.usedmarket.presenter.MarketUsedContract;
 import in.koreatech.koin.core.recyclerview.RecyclerClickListener;
 import in.koreatech.koin.core.recyclerview.RecyclerViewClickListener;
 import in.koreatech.koin.core.swiperefreshbottom.SwipeRefreshLayoutBottom;
+import in.koreatech.koin.core.toast.ToastUtil;
 import in.koreatech.koin.data.network.entity.Item;
 import in.koreatech.koin.data.network.interactor.MarketUsedRestInteractor;
 import in.koreatech.koin.data.network.response.MarketPageResponse;
-import in.koreatech.koin.ui.usedmarket.presenter.MarketUsedPresenter;
-import in.koreatech.koin.core.toast.ToastUtil;
+import in.koreatech.koin.ui.koinfragment.KoinBaseFragment;
+import in.koreatech.koin.ui.main.MainActivity;
 import in.koreatech.koin.ui.usedmarket.adapter.MarketUsedSellRecyclerAdapter;
+import in.koreatech.koin.ui.usedmarket.presenter.MarketUsedContract;
+import in.koreatech.koin.ui.usedmarket.presenter.MarketUsedPresenter;
+import in.koreatech.koin.util.NavigationManger;
 
-import java.util.ArrayList;
-
-public class MarketUsedSellFragment extends MarketUsedBaseFragment implements MarketUsedContract.View, SwipeRefreshLayoutBottom.OnRefreshListener {
+public class MarketUsedSellFragment extends KoinBaseFragment implements MarketUsedContract.View, SwipeRefreshLayoutBottom.OnRefreshListener {
     private final String TAG = "MarketUsedSellFragment";
-
-
+    private final int SELLMARKETID = 0;
+    @BindView(R.id.market_used_sell_recyclerview)
+    RecyclerView marketSellRecyclerView;
+    @BindView(R.id.market_used_sell_swiperefreshlayout)
+    SwipeRefreshLayoutBottom marketSellSwipeRefreshLayout;
     private Unbinder unbinder;
     private MarketUsedPresenter marketUsedPresenter;
     private ArrayList<Item> marketSellArrayList = new ArrayList<>(); //DB의 중고장터 리스트를 저장할 ArrayList
     /* Adapter */
     private MarketUsedSellRecyclerAdapter marketUsedSellRecyclerAdapter;
-
     /* View Component */
     private View view;
     private RecyclerView.LayoutManager layoutManager; // RecyclerView LayoutManager
-
-    private final int SELLMARKETID = 0;
-    private int postion;
+    private int position;
     private boolean grantCheck;
     private int currentPage;
     private int totalPage;
     private boolean isResume;
+    //recyclerview item touch listener
+    private RecyclerClickListener recyclerItemtouchListener = new RecyclerClickListener(getActivity(), marketSellRecyclerView, new RecyclerViewClickListener() {
+        @Override
+        public void onClick(View view, final int index) {
+            marketUsedPresenter.readGrantedDetail(marketSellArrayList.get(index).getId());
+            position = index;
 
+        }
 
-    @BindView(R.id.market_used_sell_recyclerview)
-    RecyclerView marketSellRecyclerView;
-    @BindView(R.id.market_used_sell_swiperefreshlayout)
-    SwipeRefreshLayoutBottom marketSellSwipeRefreshLayout;
+        @Override
+        public void onLongClick(View view, int position) {
+        }
+
+    });
 
     @Override
     public void onAttach(Context context) {
@@ -127,33 +136,18 @@ public class MarketUsedSellFragment extends MarketUsedBaseFragment implements Ma
 
     @Override
     public void showLoading() {
-        ((MarketUsedActivity) getActivity()).showProgressDialog(R.string.loading);
+        ((MainActivity) getActivity()).showProgressDialog(R.string.loading);
     }
 
     @Override
     public void hideLoading() {
-        ((MarketUsedActivity) getActivity()).hideProgressDialog();
+        ((MainActivity) getActivity()).hideProgressDialog();
     }
 
     public void updateUserInterface() {
 
         marketUsedSellRecyclerAdapter.notifyDataSetChanged();
     }
-
-    //recyclerview item touch listener
-    private RecyclerClickListener recyclerItemtouchListener = new RecyclerClickListener(getActivity(), marketSellRecyclerView, new RecyclerViewClickListener() {
-        @Override
-        public void onClick(View view, final int index) {
-            marketUsedPresenter.readGrantedDetail(marketSellArrayList.get(index).getId());
-            postion = index;
-
-        }
-
-        @Override
-        public void onLongClick(View view, int position) {
-        }
-
-    });
 
     @Override
     public void onRefresh() {
@@ -214,12 +208,12 @@ public class MarketUsedSellFragment extends MarketUsedBaseFragment implements Ma
 
     @Override
     public void onGrantedDataReceived(boolean granted) {
-        grantCheck = granted;
-        Intent intent = new Intent(getActivity(), MarketUsedSellDetailActivity.class);
-        intent.putExtra("ITEM_ID", marketSellArrayList.get(postion).getId());
-        intent.putExtra("MARKET_ID", SELLMARKETID);
-        intent.putExtra("GRANT_CHECK", grantCheck);
-        startActivity(intent);
+        this.grantCheck = granted;
+        Bundle bundle = new Bundle();
+        bundle.putInt("ITEM_ID", marketSellArrayList.get(position).getId());
+        bundle.putInt("MARKET_ID", SELLMARKETID);
+        bundle.putBoolean("GRANT_CHECK", this.grantCheck);
+        NavigationManger.getNavigationController(getActivity()).navigate(R.id.navi_market_used_buy_detail_action, bundle, NavigationManger.getNavigationAnimation());
     }
 
     @Override
