+++ koin/src/main/res/layout/market_used_fragment_buy.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
               android:layout_width="match_parent"
               android:layout_height="match_parent">
 
@@ -27,4 +27,4 @@
         </in.koreatech.koin.core.swiperefreshbottom.SwipeRefreshLayoutBottom>
 
     </RelativeLayout>
-</LinearLayout>
+</FrameLayout>
