+++ koin/src/main/res/layout/market_used_fragment_main.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/white"
+    android:fitsSystemWindows="true">
+
+    <in.koreatech.koin.core.appbar.AppBarBase
+        android:id="@+id/koin_base_app_bar_dark"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:leftButtonBackground="@drawable/ic_back_arrow"
+        app:rightButtonBackground="@drawable/ic_create"
+        app:titleText="중고장터" />
+
+    <androidx.viewpager.widget.ViewPager
+        android:id="@+id/market_used_main_viewpager"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="56dp"
+        android:background="@android:color/white">
+
+        <com.google.android.material.tabs.TabLayout
+            android:id="@+id/market_used_main_tabs"
+            android:layout_width="match_parent"
+            android:layout_height="47dp"
+            android:layout_gravity="top"
+            app:tabIndicatorColor="@color/colorAccent"
+            app:tabSelectedTextColor="@color/black"
+            app:tabTextColor="@color/gray9" />
+
+    </androidx.viewpager.widget.ViewPager>
+
+</FrameLayout>
+
